+++ WargameEngine/WargameEngine/WargameEngine/controller/CommandMoveObject.h
@@ -9,14 +9,14 @@ class CCommandMoveObject :
 	public ICommand
 {
 public:
-	CCommandMoveObject(std::shared_ptr<IObject> object, double deltaX, double deltaY);
+	CCommandMoveObject(std::shared_ptr<IObject> object, float deltaX, float deltaY);
 	CCommandMoveObject(IReadMemoryStream & stream, IGameModel& model);
 	void Execute();
 	void Rollback();
 	void Serialize(IWriteMemoryStream & stream) const;
 private:
 	std::shared_ptr<IObject> m_pObject;
-	double m_deltaX;
-	double m_deltaY;
+	float m_deltaX;
+	float m_deltaY;
 };
